gpMgmt unittests: Enable subtest context manager in unittest
python 3.4 and up supports subtest context manager which can be used to
dynamically create unit tests. This is really helpful in cases where only the
expectation changes based on the test input and the boiler plate code to test
for assertions and such remains the same.

When calling a subtest, a message can be specified which will show up as the
third argument in the test description. This commit changes the code so that if
splitting the test description results in three values, we add the third value
to the test details header under the "Sub Test Name" section.

Co-authored-by: David Krieger <dkrieger@vmware.com>